+++ packages/layouts/bolt-holy-grail/index.js
@@ -1,1 +1,11 @@
-import(/* webpackChunkName: 'bolt-holy-grail' */ './src/holy-grail.js');
+const holygrails = document.querySelectorAll('.l-bolt-holy-grail');
+
+if (holygrails.length) {
+  import(/* webpackChunkName: 'bolt-holy-grail' */ './src/holy-grail').then(
+    ({ BoltHolyGrail }) => {
+      holygrails.forEach(el => {
+        const holyGrailComponent = new BoltHolyGrail(el);
+      });
+    },
+  );
+}