+++ packages/testing/testing-jest/jest-global-teardown.js
@@ -1,6 +1,7 @@
 const os = require('os');
 const rimraf = require('rimraf');
 const path = require('path');
+const internalTasks = require('@bolt/build-tools/tasks/internal-tasks');
 
 const { teardown: teardownDevServer } = require('jest-dev-server');
 const { getConfig } = require('@bolt/build-tools/utils/config-store');
@@ -21,6 +22,7 @@ module.exports = async function() {
     item => !item.includes('__tests__/fixtures'),
   );
 
+  await internalTasks.clean('packages/generators/tmp'); // cleaning bolt-generator temp files after error
   await iconTasks.build(); // cleaning icons after all tests
   await teardownDevServer();
   await buildPrep(true); // clear out all built www folders when complete